Make selection of brewer more generic
