+++ app/src/main/java/views/activities/BrewingActivity.java
@@ -1,6 +1,5 @@
 package views.activities;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -9,20 +8,15 @@ import android.widget.Button;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
 import com.google.gson.Gson;
 
 import java.math.RoundingMode;
 import java.text.DecimalFormat;
-import java.util.Random;
 
 import dk.dtu.gruppeb3.broeg.app.R;
 import models.Brew;
-import models.Brewer;
 
-public class BrewingActivity extends AppCompatActivity implements View.OnClickListener {
+public class BrewingActivity extends BrewerActivity implements View.OnClickListener {
 
     public static final String SELECTED_BREW_IDENTIFIER = "selectedBrew";
 
@@ -34,8 +28,6 @@ public class BrewingActivity extends AppCompatActivity implements View.OnClickLi
     private TextView minutesLeftView;
     private TextView currentRecipeView;
 
-    private int SELECT_BREWER_REQUEST_CODE = (new Random()).nextInt(16);
-    private Brewer brewer;
     private Brew brew;
     private int totalTime;
 
@@ -61,8 +53,6 @@ public class BrewingActivity extends AppCompatActivity implements View.OnClickLi
         this.totalTime = this.brew.getTotalBrewTime();
         minutesLeftView.setText(getString(R.string.time_in_minutes, String.valueOf(Math.round((this.totalTime) / 60))));
         currentRecipeView.setText(getString(R.string.current_recipe, this.brew.getName()));
-
-        getBrewer();
     }
 
     private void startBrewing() {
@@ -113,25 +103,6 @@ public class BrewingActivity extends AppCompatActivity implements View.OnClickLi
     private void getBrew() {
         String json = getIntent().getStringExtra(SELECTED_BREW_IDENTIFIER);
         this.brew = gson.fromJson(json, Brew.class);
-    }
-
-    private void getBrewer() {
-        Intent i = new Intent(this, SelectBrewerActivity.class);
-        startActivityForResult(i, SELECT_BREWER_REQUEST_CODE);
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (requestCode == SELECT_BREWER_REQUEST_CODE && resultCode == RESULT_OK) {
-            String brewerJson = data.getStringExtra(SelectBrewerActivity.SELECTED_BREWER_IDENTIFIER);
-
-            if (brewerJson.equals(""))
-                return;
-
-            this.brewer = gson.fromJson(brewerJson, Brewer.class);
-        }
     }
 
     @Override
